Ids zu Type long konvertiert

+++ src/main/java/de/hhu/propra/exambyte/application/services/repository/NutzerInRepository.java
@@ -6,12 +6,12 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface NutzerInRepository extends CrudRepository<NutzerIn, Integer> {
+public interface NutzerInRepository extends CrudRepository<NutzerIn, Long> {
     NutzerIn save(NutzerIn nutzerIn);
 
     List<NutzerIn> findAll();
 
-    Optional<NutzerIn> findById(Integer id);
+    Optional<NutzerIn> findById(Long id);
 
     Optional<NutzerIn> findByUsername(String username);
 }
